remove legacy BuyNowArtworksRail
